Mini-Project-4: FrontEnd Integration to Blog Application
Objective:
The objective of this mini-project is to enhance the basic blog web application that you created using Node.js, Express.js, EJS as part of the Mini-Project-1. The app will enable users to create, view, edit, and delete blog posts. Previously, you were using server-side rendering using EJS files. In this project, you will be integrating React.js frontend to your app, replacing the EJS files.

You can build this mini-project either on top of Mini-Project-1, or Mini-Project-3. That is, having database in the backend is not mandatory; you have the freedom to choose between Database and No Database implementaion.